+++ changelog.md
@@ -17,6 +17,7 @@ fix:
 - caught exception in print_yield to handle Cut with samplePattern
 - corrected print yield handling of zero events after cuts
 - fixed simoultaneous log and density th2 pyplots
+- fixed verbosity option
 
 cmgrdf commit bump
 - Plots cleaned, just TH object saved in TFiles